Refactor: #215 config 패키지를 common으로 이동
